acceptance: centralize the use of build tags
This allows linters to analyze this package like any other, while still
preventing `make test` from running its tests in the absence of the
acceptance build tag.
